separating get_logger into utils
